Move code from headers to C file
