+++ src/controller.C
@@ -1,7 +1,5 @@
 #include "controller.H"
-#include <math.h>
-#include <stdio.h>
-#include <stdlib.h>
+
 
 controller::controller() {
 }
@@ -12,33 +10,16 @@ controller::controller(const controller& orig) {
 controller::~controller() {
 }
 
-void controller::checkUserInput() {
-	int userInputInt;
-	double userInputDouble;
-	while(true) {
-		currentMenu.printOptionsList();
-		// printf("\nEnter your option:\n");
-		// scanf("%lf", &userInputDouble);
-		// userInputInt = (int)floor(userInputDouble);
-
-		while(userInputInt < 0 || userInputInt > currentMenu.getNumOptions()) {
-			printf("Please enter a number between 0 and %d.", currentMenu.getNumOptions());
-			//scanf("%lf", &userInputDouble);
-			cin >> userInputDouble;
-			while(cin.fail()) {
-				cin.clear();
-				std::cin.ignore(std::numeric_limits<std::streamsize>::max(),'\n');
-				printf("Enter a number");
-			}
-			userInputInt = (int)floor(userInputDouble);
-		}
-		if(userInputInt == 0) {
-			printf("Invalid Menu Number and Option combination. Exiting program.\n");
-			exit(-1);
-		}
-		else {
-			previousMenu = currentMenu;
-			this->setCurrentMenu(userInputInt);
-		}
+void controller::setMenuList(vector<menu*>* menuList) {
+	for(int i = 0; i < menuList->size(); i++) {
+		this->menuList.push_back(menuList->at(i));
 	}
+	this->setCurrentMenu(0);
+}
+
+void controller::setCurrentMenu(int option) { currentMenu = *this->getMenu(option); }
+
+// TODO: This
+void controller::checkUserInput() {
+
 }
