+++ src/init.H
@@ -1,14 +1,6 @@
 #ifndef INIT_H
 #define INIT_H
 
-#include "menu.H"
-#include "option.H"
-#include "selectable.H"
-#include "utils.H"
-#include "viewable.H"
-
-#include <vector>
-
 using namespace std;
 
 #if defined (__cplusplus)
