+++ src/menu.C
@@ -22,17 +22,25 @@ menu::menu(const menu& orig) {
 menu::~menu() {
 }
 
+void printOptionsList() {
+    for(int i = 1; i < options.size(); i++) {
+        cout << options.at(i)->getText();
+    }
+
+    cout << options[0]->getText();
+}
+
 void menu::toString(){
 
-    //int cnt = 0;
-    //cout << this->title << endl;
-    //for(auto i : this->options)
-    //{
-    //    cout << ++cnt << ". "
-    //    i->toString();
-    //    cout << endl;
-    //}
-    //cout << endl;
+    int cnt = 0;
+    cout << this->title << endl;
+    for(auto i : this->options)
+    {
+       cout << ++cnt << ". "
+       i->toString();
+       cout << endl;
+    }
+    cout << endl;
 
 }
